+++ lerna.json
@@ -3,6 +3,6 @@
     "packages/*"
   ],
   "useWorkspaces": true,
-  "version": "0.1.0",
+  "version": "independent",
   "npmClient": "yarn"
 }
