+++ package.json
@@ -8,13 +8,10 @@
     "./packages/*"
   ],
   "scripts": {
-    "test": "jest",
-    "test:ci": "CI=1 npm run test",
+    "test": "lerna run test",
+    "test:ci": "CI=1 npm test",
     "lint": "lerna run --parallel lint",
-    "start": "if-env NODE_ENV=production && npm run -s serve || npm run -s dev",
     "build": "lerna run build",
-    "serve": "npm run build && preact serve",
-    "dev": "preact watch --template src/index.ejs",
     "install": "lerna bootstrap"
   },
   "keywords": [],
@@ -26,63 +23,6 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "@types/jest": "^23.1.1",
-    "awesome-typescript-loader": "^3.5.0",
-    "eslint": "^5.15.1",
-    "identity-obj-proxy": "^3.0.0",
-    "if-env": "^1.0.4",
-    "jest": "^21.2.1",
-    "lerna": "^3.13.1",
-    "preact-cli": "^2.0.2",
-    "preact-cli-plugin-typescript": "^0.2.2",
-    "preact-render-to-string": "^3.7.0",
-    "preact-widget-scripts": "^0.6.0",
-    "ts-jest": "^22.4.6",
-    "typescript": "^3.3.3333"
-  },
-  "dependencies": {
-    "grid-styled": "^4.1.1",
-    "preact": "^8.4.2",
-    "preact-compat": "^3.18.4",
-    "rebass": "^2.0.0-7",
-    "yup": "^0.26.10"
-  },
-  "jest": {
-    "verbose": true,
-    "transform": {
-      "^.+\\.(ts|tsx)$": "ts-jest"
-    },
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js"
-    ],
-    "testRegex": "/packages/.*/src/.*\\.test\\.(ts|tsx)$",
-    "testEnvironment": "jsdom",
-    "moduleNameMapper": {
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
-      "\\.(css|scss|less)$": "identity-obj-proxy",
-      "^react$": "preact-compat",
-      "^react-dom$": "preact-compat",
-      "@/(.*)$": "<rootDir>/src/$1"
-    },
-    "coverageThreshold": {
-      "global": {
-        "branches": 100,
-        "functions": 100,
-        "lines": 100,
-        "statements": 100
-      }
-    },
-    "modulePathIgnorePatterns": [
-      "<root>/build",
-      "<root>/packages/*/build"
-    ],
-    "moduleDirectories": [
-      "node_modules",
-      "packages",
-      "src"
-    ],
-    "setupTestFrameworkScriptFile": "<rootDir>/src/setupTests.js"
+    "lerna": "^3.13.1"
   }
 }
