+++ packages/perry-components/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "test": "jest",
     "build": "tsc",
-    "lint": "tslint -c tslint.json -p tsconfig.json"
+    "lint": "tslint --project ."
   },
   "repository": {
     "type": "git",
@@ -16,6 +16,7 @@
   "author": "perry",
   "license": "MIT",
   "devDependencies": {
+    "@types/jest": "^24.0.11",
     "jest": "^24.5.0",
     "preact-render-to-string": "^4.1.0",
     "ts-jest": "^24.0.0",
@@ -35,6 +36,9 @@
     "moduleNameMapper": {
       "^react$": "preact-compat",
       "^react-dom$": "preact-compat"
-    }
+    },
+    "modulePaths": [
+      "src"
+    ]
   }
 }
