+++ packages/perry-interfaces/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 0",
     "build": "tsc",
-    "lint": "tslint -c tslint.json -p tsconfig.json"
+    "lint": "tslint --project ."
   },
   "repository": {
     "type": "git",
